+++ resources/views/components/tools/toolbar/bootstrap.blade.php
@@ -220,8 +220,8 @@
                             <a
                                 href="#"
                                 wire:click="{{ $action }}"
-                                @if(in_array($action, $component->getBulkActionConfirms()))
-                                        wire:confirm="{{ $component->getBulkActionConfirmMessage($action) }}"
+                                @if($component->hasConfirmationMessage($action))
+                                    wire:confirm="{{ $component->getBulkActionConfirmMessage($action) }}"
                                 @endif
                                 wire:key="{{ $tableName }}-bulk-action-{{ $action }}"
                                 @class([